using a project_scoped concern
