+++ engines/dradis-api/app/controllers/dradis/ce/api/v1/project_scoped_controller.rb
@@ -10,17 +10,7 @@ module Dradis::CE::API
     # the server every time you make a change.)
     class ProjectScopedController < Dradis::CE::API::APIController
       include ActivityTracking
-
-      before_action :set_project
-      helper_method :current_project
-
-      def set_project
-        current_project
-      end
-
-      def current_project
-        @current_project ||= Project.new
-      end
+      include Dradis::CE::API::ProjectScoped
     end
   end
 end
